+++ src/main/java/svenhjol/charmony/tweaks/client/features/grindstone_disenchanting/Common.java
@@ -1,6 +1,6 @@
 package svenhjol.charmony.tweaks.client.features.grindstone_disenchanting;
 
-import svenhjol.charmony.tweaks.TweaksMod;
+import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.tweaks.common.features.grindstone_disenchanting.GrindstoneDisenchanting;
 import svenhjol.charmony.tweaks.common.features.grindstone_disenchanting.Handlers;
 
@@ -8,7 +8,7 @@ public final class Common {
     public final Handlers handlers;
 
     public Common() {
-        var common = TweaksMod.instance().sidedFeature(GrindstoneDisenchanting.class);
+        var common = Mod.getSidedFeature(GrindstoneDisenchanting.class);
         handlers = common.handlers;
     }
 }