+++ src/main/java/svenhjol/charmony/tweaks/client/features/item_frame_hiding/Common.java
@@ -1,6 +1,6 @@
 package svenhjol.charmony.tweaks.client.features.item_frame_hiding;
 
-import svenhjol.charmony.tweaks.TweaksMod;
+import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.tweaks.common.features.item_frame_hiding.ItemFrameHiding;
 import svenhjol.charmony.tweaks.common.features.item_frame_hiding.Registers;
 
@@ -8,7 +8,7 @@ public final class Common {
     public final Registers registers;
 
     public Common() {
-        var common = TweaksMod.instance().sidedFeature(ItemFrameHiding.class);
+        var common = Mod.getSidedFeature(ItemFrameHiding.class);
         registers = common.registers;
     }
 }