+++ src/main/java/svenhjol/charmony/tweaks/client/features/mob_textures/custom_renderers/CustomRenderer.java
@@ -1,11 +1,11 @@
 package svenhjol.charmony.tweaks.client.features.mob_textures.custom_renderers;
 
-import svenhjol.charmony.tweaks.TweaksMod;
+import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.tweaks.client.features.mob_textures.Handlers;
 import svenhjol.charmony.tweaks.client.features.mob_textures.MobTextures;
 import svenhjol.charmony.tweaks.client.features.mob_textures.Registers;
 
 public interface CustomRenderer {
-    Handlers handlers = TweaksMod.instance().sidedFeature(MobTextures.class).handlers;
-    Registers registers = TweaksMod.instance().sidedFeature(MobTextures.class).registers;
+    Handlers handlers = Mod.getSidedFeature(MobTextures.class).handlers;
+    Registers registers = Mod.getSidedFeature(MobTextures.class).registers;
 }