+++ src/main/java/svenhjol/charmony/tweaks/client/features/piglin_pointing/PiglinPointing.java
@@ -4,11 +4,10 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 import java.util.function.Supplier;
 
-@FeatureDefinition(side = Side.Client, showInConfig = false)
+@FeatureDefinition(side = Side.Client, canBeDisabledInConfig = false)
 public final class PiglinPointing extends SidedFeature {
     public final Handlers handlers;
     public final Supplier<Common> common;
@@ -21,6 +20,6 @@ public final class PiglinPointing extends SidedFeature {
     }
 
     public static PiglinPointing feature() {
-        return TweaksMod.instance().sidedFeature(PiglinPointing.class);
+        return Mod.getSidedFeature(PiglinPointing.class);
     }
 }