+++ src/main/java/svenhjol/charmony/tweaks/client/features/shields_have_reduced_view_blocking/ShieldsHaveReducedViewBlocking.java
@@ -6,7 +6,6 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 @FeatureDefinition(side = Side.Client, description = """
     Reduces the size and position of the shield to prevent blocking the player's view.""")
@@ -33,7 +32,7 @@ public final class ShieldsHaveReducedViewBlocking extends SidedFeature {
     }
 
     public static ShieldsHaveReducedViewBlocking feature() {
-        return TweaksMod.instance().sidedFeature(ShieldsHaveReducedViewBlocking.class);
+        return Mod.getSidedFeature(ShieldsHaveReducedViewBlocking.class);
     }
 
     public double verticalPosition() {