+++ src/main/java/svenhjol/charmony/tweaks/common/features/deepslate_dungeons/DeepslateDungeons.java
@@ -4,7 +4,6 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 @FeatureDefinition(side = Side.Common, description = """
     Dungeons in the deepslate layer will be constructed of deepslate bricks and cobbled deepslate.""")
@@ -17,6 +16,6 @@ public final class DeepslateDungeons extends SidedFeature {
     }
 
     public static DeepslateDungeons feature() {
-        return TweaksMod.instance().sidedFeature(DeepslateDungeons.class);
+        return Mod.getSidedFeature(DeepslateDungeons.class);
     }
 }