+++ src/main/java/svenhjol/charmony/tweaks/common/features/parrots_stay_on_shoulder/ParrotsStayOnShoulder.java
@@ -4,7 +4,6 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 @FeatureDefinition(side = Side.Common, description = """
     Parrots stay on player's shoulder when jumping and falling. Crouch to make them dismount.""")
@@ -22,6 +21,6 @@ public final class ParrotsStayOnShoulder extends SidedFeature {
     }
 
     public static ParrotsStayOnShoulder feature() {
-        return TweaksMod.instance().sidedFeature(ParrotsStayOnShoulder.class);
+        return Mod.getSidedFeature(ParrotsStayOnShoulder.class);
     }
 }