+++ src/main/java/svenhjol/charmony/tweaks/common/features/piglin_pointing/PiglinPointing.java
@@ -4,7 +4,6 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 @FeatureDefinition(side = Side.Common, description = """
     Piglins turn and point in the rough direction of a nether fortress or bastion remnant.
@@ -24,6 +23,6 @@ public final class PiglinPointing extends SidedFeature {
     }
 
     public static PiglinPointing feature() {
-        return TweaksMod.instance().sidedFeature(PiglinPointing.class);
+        return Mod.getSidedFeature(PiglinPointing.class);
     }
 }