+++ src/main/java/svenhjol/charmony/tweaks/common/features/pigs_find_mushrooms/PigsFindMushrooms.java
@@ -6,7 +6,6 @@ import svenhjol.charmony.core.annotations.FeatureDefinition;
 import svenhjol.charmony.core.base.Mod;
 import svenhjol.charmony.core.base.SidedFeature;
 import svenhjol.charmony.core.enums.Side;
-import svenhjol.charmony.tweaks.TweaksMod;
 
 @FeatureDefinition(side = Side.Common, description = """
     Pigs have a chance to find mushrooms from certain blocks. By default mushrooms are unearthed from mycelium and podzol blocks.
@@ -39,7 +38,7 @@ public final class PigsFindMushrooms extends SidedFeature {
     }
 
     public static PigsFindMushrooms feature() {
-        return TweaksMod.instance().sidedFeature(PigsFindMushrooms.class);
+        return Mod.getSidedFeature(PigsFindMushrooms.class);
     }
 
     public int findChance() {